Revert "gtv3 variants"
