+++ .gitignore
@@ -1,6 +1,5 @@
 lib/__pycache__/*
 models/__pycache__/*
-models/gtv3/__pycache__/*
 task2_digress_qm9.ipynb
 models/digress.py
 models/gt_digress.py
