+++ models/gtv2_gated.py
@@ -1,6 +1,5 @@
 import torch
 import torch.nn as nn
-from models.attention import *
 
 
 def sym_tensor(x):
@@ -12,23 +11,90 @@ def sym_tensor(x):
     return x               # [bs, n, n, d]
 
 
+class attention_node_to_edge(nn.Module):
+    def __init__(self, d, d_head, drop=0.0):
+        super().__init__()
+        self.Q_node = nn.Linear(d, d_head)
+        self.K_edge = nn.Linear(d, d_head)
+        self.V_edge = nn.Linear(d, d_head)
+        self.sqrt_d = torch.sqrt(torch.tensor(d_head))
+        self.drop_att = nn.Dropout(drop)
+    def forward(self, x, e):
+        Q_node = self.Q_node(x) # [bs, n, d_head]
+        K_edge = self.K_edge(e) # [bs, n, n, d_head]
+        V_edge = self.V_edge(e) # [bs, n, n, d_head]
+        Q_node = Q_node.unsqueeze(2) # [bs, n, 1, d_head]
+        Att = (Q_node * K_edge).sum(dim=3) / self.sqrt_d # [bs, n, n]
+        Att = torch.softmax(Att, dim=2)
+        Att = self.drop_att(Att)
+        Att = Att.unsqueeze(-1) # [bs, n, n, 1]
+        x = (Att * V_edge).sum(dim=2) # [bs, n, d_head]
+        return x, e # [bs, n, d_head]
+
+
+class attention_edge_to_node(nn.Module):
+    def __init__(self, d, d_head, drop=0.0):
+        super().__init__()
+        self.Q_edge = nn.Linear(d, d_head)
+        self.K_node = nn.Linear(d, d_head)
+        self.V_node = nn.Linear(d, d_head)
+        self.sqrt_d = torch.sqrt(torch.tensor(d_head))
+        self.drop_att = nn.Dropout(drop)
+    def forward(self, x, e):
+        Q_edge = self.Q_edge(e) # [bs, n, n, d_head]
+        K_node = self.K_node(x) # [bs, n, d_head]
+        V_node = self.V_node(x) # [bs, n, d_head]
+        K_i = K_node.unsqueeze(1).expand(-1, e.size(1), -1, -1) # [bs, n, n, d_head]
+        V_i = V_node.unsqueeze(1).expand(-1, e.size(1), -1, -1) # [bs, n, n, d_head]
+        K_j = K_node.unsqueeze(2).expand(-1, -1, e.size(1), -1) # [bs, n, n, d_head]
+        V_j = V_node.unsqueeze(2).expand(-1, -1, e.size(1), -1) # [bs, n, n, d_head]
+        Att_i = torch.exp((Q_edge * K_i).sum(dim=-1) / self.sqrt_d) # [bs, n, n]
+        Att_j = torch.exp((Q_edge * K_j).sum(dim=-1) / self.sqrt_d) # [bs, n, n]
+        Att_sum = Att_i + Att_j
+        Att_i = self.drop_att(Att_i / Att_sum)
+        Att_j = self.drop_att(Att_j / Att_sum)
+        e = Att_i.unsqueeze(-1) * V_i + Att_j.unsqueeze(-1) * V_j # [bs, n, n, d_head]
+        return x, e
+
+
+class attention_node_to_node(nn.Module):
+    def __init__(self, d, d_head, drop=0.0):
+        super().__init__()
+        self.Q = nn.Linear(d, d_head)  # For node queries
+        self.K = nn.Linear(d, d_head)  # For node keys
+        self.V = nn.Linear(d, d_head)  # For node values
+        self.sqrt_d = torch.sqrt(torch.tensor(d_head))
+        self.dropout = nn.Dropout(drop)
+    def forward(self, x, e):
+        Q = self.Q(x)  # [bs, n, d_head]
+        K = self.K(x)  # [bs, n, d_head]
+        V = self.V(x)  # [bs, n, d_head]
+        Q = Q.unsqueeze(2)  # [bs, n, 1, d_head]
+        K = K.unsqueeze(1)  # [bs, 1, n, d_head]
+        Att = (Q * K).sum(dim=-1) / self.sqrt_d  # [bs, n, n]
+        Att = torch.softmax(Att, dim=-1)  # [bs, n, n]
+        Att = self.dropout(Att)
+        x = Att @ V  # [bs, n, d_head]
+        return x, e
+
+
 class head_attention_gated(nn.Module):
     def __init__(self, d, d_head, drop):
         super().__init__()
-        self.cross_att_edge_to_node = attention_edge_to_node(d, d_head, drop)
         self.cross_att_node_to_edge = attention_node_to_edge(d, d_head, drop)
-        self.self_att_node_to_node = attention_node_to_node(d, d_head, drop)
+        self.cross_att_edge_to_node = attention_edge_to_node(d, d_head, drop)
+        self.cross_att_node_to_node = attention_node_to_node(d, d_head, drop)
         self.gated_mlp = nn.Sequential(
             nn.Linear(2 * d_head, d_head),
             nn.Sigmoid(),
         )
     def forward(self, x, e):
-        # 1) Cross-attention from edges to nodes
-        x_cross, _ = self.cross_att_edge_to_node(x, e)
+        # 1) Cross-attention from nodes to edges
+        x_cross, _ = self.cross_att_node_to_edge(x, e)
         # 2) Cross-attention from edges to nodes
-        _, e = self.cross_att_node_to_edge(x, e)
+        _, e = self.cross_att_edge_to_node(x, e)
         # 3) Self-attention on nodes
-        x_self, _ = self.self_att_node_to_node(x, e)
+        x_self, _ = self.cross_att_node_to_node(x, e)
         # 4) Gated MLP
         g = self.gated_mlp(torch.cat([x_cross, x_self], dim=-1))
         x = g * x_cross + (1 - g) * x_self
