--- models/gtv3/sa__ca_e_h__ca_n_h_e.py
@@ -1,119 +1,0 @@
-import torch
-import torch.nn as nn
-from models.attention import attention_node_to_node
-
-
-def sym_tensor(x):
-    x = x.permute(0,3,1,2) # [bs, d, n, n]
-    triu = torch.triu(x,diagonal=1).transpose(3,2) # [bs, d, n, n]
-    mask = (triu.abs()>0).float()                  # [bs, d, n, n]
-    x =  x * (1 - mask ) + mask * triu             # [bs, d, n, n]
-    x = x.permute(0,2,3,1) # [bs, n, n, d]
-    return x               # [bs, n, n, d]
-
-
-class attention_edge_to_node(nn.Module):
-    def __init__(self, d, d_head, drop=0.0):
-        super().__init__()
-        self.Q_node = nn.Linear(d_head, d_head)
-        self.K_edge = nn.Linear(d_head, d_head)
-        self.V_edge = nn.Linear(d_head, d_head)
-        self.sqrt_d = torch.sqrt(torch.tensor(d_head))
-        self.drop_att = nn.Dropout(drop)
-    def forward(self, x, e):
-        Q_node = self.Q_node(x) # [bs, n, d_head]
-        K_edge = self.K_edge(e) # [bs, n, n, d_head]
-        V_edge = self.V_edge(e) # [bs, n, n, d_head]
-        Q_node = Q_node.unsqueeze(2) # [bs, n, 1, d_head]
-        Att = (Q_node * K_edge).sum(dim=3) / self.sqrt_d # [bs, n, n]
-        Att = torch.softmax(Att, dim=2)
-        Att = self.drop_att(Att)
-        Att = Att.unsqueeze(-1) # [bs, n, n, 1]
-        x = (Att * V_edge).sum(dim=2) # [bs, n, d_head]
-        return x, e # [bs, n, d_head]
-
-
-class attention_node_to_edge(nn.Module):
-    def __init__(self, d, d_head, drop=0.0):
-        super().__init__()
-        self.Q_edge = nn.Linear(d, d_head)
-        self.K_node = nn.Linear(d_head, d_head)
-        self.V_node = nn.Linear(d_head, d_head)
-        self.sqrt_d = torch.sqrt(torch.tensor(d_head))
-        self.drop_att = nn.Dropout(drop)
-    def forward(self, x, e):
-        Q_edge = self.Q_edge(e) # [bs, n, n, d_head]
-        K_node = self.K_node(x) # [bs, n, d_head]
-        V_node = self.V_node(x) # [bs, n, d_head]
-        K_i = K_node.unsqueeze(1).expand(-1, e.size(1), -1, -1) # [bs, n, n, d_head]
-        V_i = V_node.unsqueeze(1).expand(-1, e.size(1), -1, -1) # [bs, n, n, d_head]
-        K_j = K_node.unsqueeze(2).expand(-1, -1, e.size(1), -1) # [bs, n, n, d_head]
-        V_j = V_node.unsqueeze(2).expand(-1, -1, e.size(1), -1) # [bs, n, n, d_head]
-        Att_i = torch.exp((Q_edge * K_i).sum(dim=-1) / self.sqrt_d) # [bs, n, n]
-        Att_j = torch.exp((Q_edge * K_j).sum(dim=-1) / self.sqrt_d) # [bs, n, n]
-        Att_sum = Att_i + Att_j
-        Att_i = self.drop_att(Att_i / Att_sum)
-        Att_j = self.drop_att(Att_j / Att_sum)
-        e = Att_i.unsqueeze(-1) * V_i + Att_j.unsqueeze(-1) * V_j # [bs, n, n, d_head]
-        return x, e
-
-
-class head_attention_v3(nn.Module):
-    def __init__(self, d, d_head, drop):
-        super().__init__()
-        self.self_att_node_to_node = attention_node_to_node(d, d_head, drop)
-        self.cross_att_node_to_edge = attention_node_to_edge(d, d_head, drop)
-        self.cross_att_edge_to_node = attention_edge_to_node(d, d_head, drop)
-    def forward(self, x, e):
-        x_hat, _ = self.self_att_node_to_node(x, e)
-        _, e_new = self.cross_att_node_to_edge(x_hat, e)
-        x_new, _ = self.cross_att_edge_to_node(x_hat, e_new)
-        return x_new, e_new
-
-
-class MHA_v3(nn.Module):
-    def __init__(self, d, num_heads, drop=0.0):  
-        super().__init__()
-        d_head = d // num_heads
-        self.heads = nn.ModuleList([head_attention_v3(d, d_head, drop) for _ in range(num_heads)])
-        self.WOx = nn.Linear(d, d)
-        self.WOe = nn.Linear(d, d)
-        self.drop_x = nn.Dropout(drop)
-        self.drop_e = nn.Dropout(drop)
-    def forward(self, x, e):
-        x_MHA = []
-        e_MHA = []    
-        for head in self.heads:
-            x_HA, e_HA = head(x,e)            # [bs, n, d_head], [bs, n, n, d_head]
-            x_MHA.append(x_HA)
-            e_MHA.append(e_HA)
-        x = self.WOx(torch.cat(x_MHA, dim=2)) # [bs, n, d]
-        x = self.drop_x(x)                    # [bs, n, d]
-        e = self.WOe(torch.cat(e_MHA, dim=3)) # [bs, n, n, d]
-        e = self.drop_e(e)                    # [bs, n, n, d]
-        return x, e                           # [bs, n, d], [bs, n, n, d]
-
-
-class BlockGT(nn.Module):
-    def __init__(self, d, num_heads, drop=0.0):  
-        super().__init__()
-        self.LNx = nn.LayerNorm(d)
-        self.LNe = nn.LayerNorm(d)
-        self.LNx2 = nn.LayerNorm(d)
-        self.LNe2 = nn.LayerNorm(d)
-        self.MHA = MHA_v3(d, num_heads, drop)
-        self.MLPx = nn.Sequential(nn.Linear(d, 4*d), nn.LeakyReLU(), nn.Linear(4*d, d))
-        self.MLPe = nn.Sequential(nn.Linear(d, 4*d), nn.LeakyReLU(), nn.Linear(4*d, d))
-        self.drop_x_mlp = nn.Dropout(drop)
-        self.drop_e_mlp = nn.Dropout(drop)
-    def forward(self, x, e):
-        x = self.LNx(x)                 # [bs, n, d]
-        e = self.LNe(e)                 # [bs, n, n, d]
-        x_MHA, e_MHA = self.MHA(x, e)   # [bs, n, d], [bs, n, n, d]
-        x = x + x_MHA                   # [bs, n, d]
-        x = x + self.MLPx(self.LNx2(x)) # [bs, n, d]
-        x = self.drop_x_mlp(x)          # [bs, n, d]
-        e = e + e_MHA                   # [bs, n, n, d]
-        e = e + self.MLPe(self.LNe2(e)) # [bs, n, n, d]
-        e = self.drop_e_mlp(e)          # [bs, n, n, d]
-        return x, e                     # [bs, n, d], [bs, n, n, d]
